implementando proteção para a exclusão de funcionários

+++ src/br/com/idecaph/server/FuncionariosServiceImpl.java
@@ -3,8 +3,13 @@ package br.com.idecaph.server;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+
 import br.com.idecaph.client.interfaces.FuncionariosService;
+import br.com.idecaph.dao.AvaliadoPesquisaDAO;
 import br.com.idecaph.dao.FuncionarioDAO;
+import br.com.idecaph.dao.ParticipantePesquisaDAO;
 import br.com.idecaph.model.Funcionario;
 import br.com.idecaph.shared.FuncionarioClient;
 
@@ -36,16 +41,45 @@ public class FuncionariosServiceImpl extends RemoteServiceServlet implements
 	}
 
 	@Override
-	public boolean excluiFuncionario(Long id) {
-		boolean ok = true;
+	public String excluiFuncionario(Long id) {
 		try {
-			FuncionarioDAO funcionarioDAO = new FuncionarioDAO();
-			funcionarioDAO.deleteById(id);
+
+			HttpServletRequest request = getThreadLocalRequest();
+
+			HttpSession session = request.getSession();
+
+			FuncionarioClient funcionarioLogado = (FuncionarioClient) session
+					.getAttribute("funcionario");
+
+			if (funcionarioLogado.getId().equals(id)) {
+				return "Não é possível excluir funcionário logado.";
+			}
+
+			ParticipantePesquisaDAO participantePesquisaDAO = new ParticipantePesquisaDAO();
+			boolean existePesquisaParticipante = participantePesquisaDAO
+					.existePesquisaFuncionario(id);
+
+			if (existePesquisaParticipante) {
+				return "Não é possível excluir esse funcionário, pois já está cadastrado em pelo menos uma pesquisa.";
+			} else {
+
+				AvaliadoPesquisaDAO avaliadoPesquisaDAO = new AvaliadoPesquisaDAO();
+				boolean existePesquisaAvaliado = avaliadoPesquisaDAO
+						.existePesquisaFuncionario(id);
+
+				if (existePesquisaAvaliado) {
+					return "Não é possível excluir esse funcionário, pois já está cadastrado em pelo menos uma pesquisa.";
+				} else {
+
+					FuncionarioDAO funcionarioDAO = new FuncionarioDAO();
+					funcionarioDAO.deleteById(id);
+				}
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
-			ok = false;
+			return "Ocorreu um erro ao excluir o funcionário. Por favor, contate o administrador do sistema.";
 		}
-		return ok;
+		return null;
 	}
 
 	@Override
